fix: add logic that upload image to s3

+++ src/main/java/com/ajou/travely/service/AwsS3Service.java
@@ -1,6 +1,7 @@
 package com.ajou.travely.service;
 
 import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -9,6 +10,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -31,6 +33,9 @@ public class AwsS3Service {
 
         // forEach 구문을 통해 multipartFile로 넘어온 파일들 하나씩 fileNameList에 추가
         multipartFile.forEach(file -> {
+            if(!Objects.requireNonNull(file.getContentType()).startsWith("image")){
+                return;
+            }
             String fileName = createFileName(file.getOriginalFilename());
             ObjectMetadata objectMetadata = new ObjectMetadata();
             objectMetadata.setContentLength(file.getSize());
@@ -50,10 +55,13 @@ public class AwsS3Service {
     }
 
     public void deleteFile(String fileName) {
+        if (!amazonS3.doesObjectExist(bucket, fileName)) {
+            throw new AmazonS3Exception("Object " + fileName + " does not exist!");
+        }
         amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
     }
 
-    private String createFileName(String fileName) { // 먼저 파일 업로드 시, 파일명을 난수화하기 위해 random으로 돌립니다.
+    private String createFileName(String fileName) { // 파일 업로드 시, 파일명을 난수화하기 위해 random 사용
         return UUID.randomUUID().toString().concat(getFileExtension(fileName));
     }
